+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.1.3] - 2022-09-13
+
+### Added
+
+- Add social media links
+
+### Changed
+
+- Update component to switch theme
+
 ## [4.1.2] - 2022-09-01
 
 ### Fixed
@@ -463,6 +473,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add RFC cross chain navigation
 - Add a switch to selcet a dark theme from navigation drawer
 
+[4.1.3]: https://github.com/commercionetwork/almerico/compare/v4.1.2...v4.1.3
+[4.1.2]: https://github.com/commercionetwork/almerico/compare/v4.1.1...v4.1.2
 [4.1.1]: https://github.com/commercionetwork/almerico/compare/v4.1.0...v4.1.1
 [4.1.0]: https://github.com/commercionetwork/almerico/compare/v4.0.0...v4.1.0
 [4.0.0]: https://github.com/commercionetwork/almerico/compare/v3.3.1...v4.0.0